Add support for the ARIES Embedded board M100PFSEVP
The M100PFS is a System On Module with the PolarFire SoC-FPGA from
Aries Embedded GmbH. The M100PFSEVP is the Evaluation Platform for
that module.

The README.md in the top-level directory describes, how to setup,
build and use the Microchip PolarFire SoC Yocto BSP. The BSP for the
M100PFSEVP can be build with:

  $ MACHINE=m100pfsevp bitbake core-image-minimal

Note: it's not possible, to define the MACHINE in "conf/local.conf"!

In the directoy "tmp-glibc/deploy/images/m100pfsevp/" the following
images are created:

  core-image-minimal-m100pfsevp.tar.xz
  core-image-minimal-m100pfsevp.wic
  fitImage
  m100pfsevp-emmc.dtb
  m100pfsevp-sdcard.dtb

The rootfs in the "tar.xz" archive can be used as NFS root on the
host. The "wic" image can be written to the eMMC or SD card.

Signed-off-by: Wolfgang Grandegger <wg@aries-embedded.de>