no do_configure_prepend, no SRC_URI_append_xxx
1) no do_configure_prepend

eSDK is unhappy with it and anyways the patch already adds .dts and .dts files

2) no SRC_URI_append_xxx

With every SRC_URI_append_ eSDK/devtool creates a branch.
This complicates things a lot and since the kerne recipe is "fragile" it does
not really work as expected.

As a workaround I removed support for other boards.